+++ app/javascript/components/carroussel.js
@@ -1,36 +1,29 @@
 import $ from 'jquery';
 
-// const moveToSelected = (element) => {
-//   if (element == "next") {
-//     var selected = $(".selected").next();
-//   } else if (element == "prev") {
-//     var selected = $(".selected").prev();
-//   } else {
-//     var selected = element;
-//   }
+function moveToSelected(element) {
+  if (element == "next") {
+    var selected = $(".selected").next();
+  } else if (element == "prev") {
+    var selected = $(".selected").prev();
+  } else {
+    var selected = element;
+  }
 
-//   var next = $(selected).next();
-//   var prev = $(selected).prev();
-//   var prevSecond = $(prev).prev();
-//   var nextSecond = $(next).next();
+  var next = $(selected).next();
+  var prev = $(selected).prev();
+  var prevSecond = $(prev).prev();
+  var nextSecond = $(next).next();
 
-//   $(selected).removeClass().addClass("selected");
+  $(selected).removeClass().addClass("selected");
 
-//   $(prev).removeClass().addClass("prev");
-//   $(next).removeClass().addClass("next");
+  $(prev).removeClass().addClass("prev");
+  $(next).removeClass().addClass("next");
 
-//   $(nextSecond).removeClass().addClass("nextRightSecond");
-//   $(prevSecond).removeClass().addClass("prevLeftSecond");
+  $(nextSecond).removeClass().addClass("nextRightSecond");
+  $(prevSecond).removeClass().addClass("prevLeftSecond");
 
-//   $(nextSecond).nextAll().removeClass().addClass('hideRight');
-//   $(prevSecond).prevAll().removeClass().addClass('hideLeft');
+  $(nextSecond).nextAll().removeClass().addClass('hideRight');
+  $(prevSecond).prevAll().removeClass().addClass('hideLeft');
+}
 
-// }
-
-
-
-// export { moveToSelected };
-
-
-
-
+export { moveToSelected }
