+++ app/javascript/packs/application.js
@@ -5,39 +5,13 @@ import { fetchWeather } from '../components/weather';
 import { initAutocomplete } from '../plugins/init_autocomplete';
 import "../plugins/flatpickr";
 import { mosaic } from '../components/mosaic';
-// import { moveToSelected } from '../components/carrousel';
+import { moveToSelected } from '../components/carroussel';
 
 initUpdateNavbarOnScroll();
 fetchWeather();
 initAutocomplete();
 mosaic();
-// moveToSelected();
 
-function moveToSelected(element) {
-  if (element == "next") {
-    var selected = $(".selected").next();
-  } else if (element == "prev") {
-    var selected = $(".selected").prev();
-  } else {
-    var selected = element;
-  }
-
-  var next = $(selected).next();
-  var prev = $(selected).prev();
-  var prevSecond = $(prev).prev();
-  var nextSecond = $(next).next();
-
-  $(selected).removeClass().addClass("selected");
-
-  $(prev).removeClass().addClass("prev");
-  $(next).removeClass().addClass("next");
-
-  $(nextSecond).removeClass().addClass("nextRightSecond");
-  $(prevSecond).removeClass().addClass("prevLeftSecond");
-
-  $(nextSecond).nextAll().removeClass().addClass('hideRight');
-  $(prevSecond).prevAll().removeClass().addClass('hideLeft');
-}
 
 // Eventos teclado
 $(document).keydown(function(e) {
@@ -67,4 +41,3 @@ $('#next').click(function() {
   moveToSelected('next');
 });
 
-// window.$ = $
